app/requirements.txt
Flask==3.0.0
flask-cors==3.0.9
elasticsearch==7.10.1
psycopg2-binary==2.9.3
pandas==2.1.0
gensim==4.1.2
gunicorn==20.0.4